Added basic test tasks and updated timer config
Updated the timebase source of the controller to use TIM10 instead of TIM1,
freeing for more important uses, and create two basic tasks that toggle pin
PC13 (LED Pin) to test the functionality of FreeRTOS on the controller.

The test was successful, the on-board LED is effectively randomly blinking.

Also contains the configuration of PC13 to a standard GPIO output.