First attempt at using an LCD screen over I2C
Followed a tutorial to interface with the PCF8574T I/O expander chip connected
to the 1602A-QAPASS lcd screen. However, upon running this program, nothing
happened, and upon inspection of the tutorial with it's likely reference
material, it is clear that certain parts are not correct.

Second attempt will follow the LCD screen library for arduino here:
https://github.com/fdebrabander/Arduino-LiquidCrystal-I2C-library

Tutorial that did not work:
https://controllerstech.com/i2c-lcd-in-stm32/

Also slightly updated the TODO list to more accurately represent current tasks.