First test of interfacing with GY-88
The test was to see if the board is still alive by seeing if it was possible to
read from the ID register and write to the configuration registers of the
MPU6050 chip on the board.

The test was successful as no error messages were sent to the serial port when
the GY-88 was connected, and the correct error messages were sent when the
GY-88 was disconnected.