Second test second attempt
The register address was double checked and is indeed correct, and after trying
several things to isolate the issue (# of bytes read, correct voltage, etc),
the HAL interface for I2C is currently stuck as HAL_BUSY, and no communication
is happening between the board and the GY-88.

Further testing is needed as the cause of this issue is unknown.