Revert "Second test of interfacing with GY-88"
This reverts commit 202eb6f44e5c7a7ffff9ad15ad0f9fd3e5dfed73.

Reverted to an older version; Testing path went in a direction unrelated to the issue

Explained the issue briefly in the previous commit, but in more detail the issue
was intially that the values from the registers of the MPU6050 were unchanging
and not at all what was expected. Then, while debugging this issue, a new issue
turned in that the MPU6050 no longer responded to its address on the i2c line.

The root cause of this new issue was that my STM32F401 and GY-88 boards were
connected to separate USB ports. I had a false assumption that the two devices
had the exact same ground. However, since the ground connected to the GY-88 was
through a long cable (relatively to the rest of the circuit), in the correct
positions it would introduce noise into the line.

I cannot say how this noise messed with the i2c communcation devices, but all I
know is that one the ground was directly connected between the two devices, the
bizarre behaviour had stopped, and the MPU6050 was responding again.

I have yet to test the precursor issue, but I can safely say that this issue has
been solved.