Began functions for getting current orientation and increased clock speed
In order to get our current position, we need to integrate the acceleration
over time. But in order to have accurate measurements, we need to check the
acceleration often and have an accurate measurement of time, so the clock speed
was increased for better time measurement.

Saving this point in case increasing the clock speed breaks something further
down the line.