Finalize gyro positioning and corrected logging files
Currently the program is able to read from the MPU6050 and produce a close
enough value of the current yaw of the device. The units are not exactly degrees
but they will be close enough for the purposes of this project.

The logging functions were also the cause of lots of undefined behaviour, so
they were corrected to prevent the processor from creating a fault by writing to
or trying to access restricted memory or similar behaviours.