Added PWM functionality for motor control at a constant speed
The output values from the encoder aren't from 0 to 360 degrees, instead they
appear to go from 50 to 100, but we need to know the exact range to be able to
create an algorithm that can determine speed from these values.

Thus we can just spin the motor slowly using PWM and look at the values that we
get.