First test of AS5600 encoder
Using PWM to drive the motor a low speed, the test was successfully able to
print the max value the of encoder as seen when used by an Arduino. However,
the min value is not the same as the Arduino, which will be the next issue that
I'll be looking into.

It also appears that I have somehow broken something in the gyro handling code,
as the values of the gyro sensor have gone bizzare.