Tested volume of encoder values at different motor speeds
This test was for helping to decide on which algorithm to use to determine
speed. When the motor is spinning slowly, there are plenty of values between the
min and max values, (from ~700 to ~1100), and when spinning at 50% speed, there
are about 10 values between min and max (with about 500 microseconds between
measurements).

However, the rate at which these values increment are not linear, so we can't
take the difference between two points to get speed, instead we'll need to
measure the time it takes to go from min to max as a measure of speed.